flexibleengine: Removed usage of package 'hashcode'
As documented on https://www.terraform.io/docs/extend/guides/v2-upgrade-guide.html#removal-of-helper-hashcode-package this package
has been removed on the v2 so we have to use a helper.

For the 'hashcode.String' I used the 'schema.HashString' as it calls the internal helper.
For the 'hashcode.Strings' I created a new helper as documented on the 'utils.go' named 'HashString'
which has the logic described